Fix errore di formattazione codice fiscale, implementazione <optional>, implementazione .stream, .map, .collect, veridfica, gestione errori id, passaggio da long a int

+++ src/main/java/com/example/bankmanapp/Controller/ContoController.java
@@ -21,7 +21,7 @@ public class ContoController {
 
 
     @GetMapping(value = "/{id}")
-    public ContoDto getConto(@PathVariable Long id){
+    public ContoDto getConto(@PathVariable int id){
         ContoDto conto = contoService.trovaPerId(id);
         return conto;
     }
