+++ src/main/java/com/example/bankmanapp/Dto/MovimentiDto.java
@@ -5,8 +5,8 @@ import java.time.LocalDateTime;
 
 public record MovimentiDto(
 
-        Long id,
-        Long idConto, // FK verso Conto
+        int id,
+        int idConto, // FK verso Conto
         BigDecimal importo,
         String tipo, // "DEPOSITO" o "PRELIEVO"
         LocalDateTime data
